e-31

+++ Assets/_Data/Item/DropRate.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class DropRate
 {
     public ItemSO itemSO;
-    public float dropRate;
+    public int dropRate;
     public int minDrop;
     public int maxDrop;
 }
